Relocate pmstats to avoid conflict with kexec hardboot page.
The last page (4 kB) of physical RAM, after the ram_console, is reserved
for pm_debug_scratchpad (pmstats).  This conflicts with that memory's usage
as the hardboot page, already in use in kexec-enabled epicmtd kernels.

Since pmstats uses very little of the whole 4 kB, move the scratchpad
forward to leave 1 kB of memory available for kexec.  Also requires fixing
up pmstats alloc and copy routines to not attempt to read the whole page.